Refactor task manager -> pkg
